miro_preprocessor/miro_image_to_tandem_vault/src/requirements.txt
daiquiri>=1.3.0
attrs>=17.2.0